# openapi/generated.yaml
components:
  schemas:
...
    ElementActionRequest:
      description: Perform a desktop action on an element.
      properties:
        action:
          description: Desktop action to perform
          enum:
          - click
          - doubleClick
          - hover
          - clickAndHold
          - contextClick
          - release
          - sendKeys
          - mouseover
          - unknown
          type: string
        elementHandle:
          description: Element handle to act on
          type: string
        input:
          description: Optional input (used by SEND_KEYS)
          nullable: true
          type: string
      required:
      - action
      - elementHandle
      type: object
...
    ElementStateResponse:
      description: Resolved element state including handle, visibility, attributes, and rect.
      nullable: true
      properties:
        attributes:
          additionalProperties:
            description: Flattened attribute map
            type: string
          description: Flattened attribute map
          type: object
        displayed:
          description: Whether the element is displayed to the user
          type: boolean
        elementHandle:
          description: Opaque handle for subsequent calls
          type: string
        found:
          description: Whether the element exists in the DOM
          type: boolean
        rect:
          $ref: '#/components/schemas/Rect'
      type: object
...
    ErrorDetail:
      description: Error detail envelope.
      properties:
        code:
          description: Machine-readable error code (stable).
          example: session_not_found
          type: string
        details:
          additionalProperties:
            description: Optional structured details keyed by field name.
            type: object
          description: Optional structured details keyed by field name.
          type: object
        message:
          description: Human-readable error message.
          type: string
        requestId:
          description: Correlating request identifier echoed back to the caller.
          type: string
      type: object
    ErrorResponse:
      description: Uniform error envelope returned for non-2xx responses.
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
    ExecuteRequest:
      description: Execute arbitrary JavaScript inside the session.
      properties:
        args:
          description: Optional positional arguments passed to the script
          items:
            description: Optional positional arguments passed to the script
            nullable: true
            type: object
          nullable: true
          type: array
        script:
          description: JavaScript source to run
          type: string
      required:
      - script
      type: object
    ExecuteResponse:
      description: Result of a JavaScript execution.
      properties:
        result:
          description: JSON-serialized result of the script
          nullable: true
          type: object
      type: object
    FindElementRequest:
      description: Locate a single element by XPath.
      properties:
        xpath:
          description: XPath expression to resolve
          type: string
      required:
      - xpath
      type: object
...
    Rect:
      description: Axis-aligned bounding rectangle in CSS pixels.
      nullable: true
      properties:
        height:
          description: Rect height
          format: int32
          type: integer
        width:
          description: Rect width
          format: int32
          type: integer
        x:
          description: Left offset
          format: int32
          type: integer
        y:
          description: Top offset
          format: int32
          type: integer
      type: object
...
    ScrollOffset:
      description: Current page scroll offset in CSS pixels.
      properties:
        x:
          description: Horizontal scroll offset
          example: 0
          format: int32
          type: integer
        y:
          description: Vertical scroll offset
          example: 1200
          format: int32
          type: integer
      type: object
...
    SessionStateResponse:
      description: Session state snapshot including current URL, viewport, and scroll offset.
      properties:
        browserType:
          description: Browser type
          enum:
          - chrome
          - firefox
          - safari
          - ie
          - android
          - ios
          type: string
        createdAt:
          description: Instant the session was created
          format: date-time
          type: string
        currentUrl:
          description: Current URL as reported by the driver
          nullable: true
          type: string
        environment:
          description: Session environment
          enum:
          - test
          - discovery
          type: string
        expiresAt:
          description: Instant at which the session will be reaped
          format: date-time
          type: string
        scrollOffset:
          $ref: '#/components/schemas/ScrollOffset'
        sessionId:
          description: Session identifier
          format: uuid
          type: string
        viewport:
          $ref: '#/components/schemas/Viewport'
      type: object
    Viewport:
      description: Browser viewport size in CSS pixels.
      properties:
        height:
          description: Viewport height (px)
          example: 720
          format: int32
          type: integer
        width:
          description: Viewport width (px)
          example: 1280
          format: int32
          type: integer
      type: object
    ViewportStateResponse:
      description: Current viewport size and scroll offset.
      properties:
        scrollOffset:
          $ref: '#/components/schemas/ScrollOffset'
        viewport:
          $ref: '#/components/schemas/Viewport'
      type: object
info:
  description: Standalone service that exposes Selenium/Appium browser sessions over HTTP.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Browser Service
  version: v1
openapi: 3.0.1
paths:
...
  /v1/sessions/{id}:
    delete:
      operationId: deleteSession
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '204':
          description: Closed
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Session not found
      summary: Close a session
      tags:
      - Sessions
    get:
      operationId: getSession
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SessionStateResponse'
          description: Session state
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Session not found
      summary: Describe a session
      tags:
      - Sessions
...
  /v1/sessions/{id}/element/action:
    post:
      operationId: performElementAction
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementActionRequest'
        required: true
      responses:
        '204':
          description: Action performed
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Session or element handle not found
        '409':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Mobile session (desktop required)
      summary: Perform a desktop action on an element
      tags:
      - Elements
  /v1/sessions/{id}/element/find:
    post:
      operationId: findElement
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FindElementRequest'
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ElementStateResponse'
          description: OK
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Session not found
      summary: Locate an element by XPath
      tags:
      - Elements
...
  /v1/sessions/{id}/execute:
    post:
      operationId: executeScript
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteRequest'
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExecuteResponse'
          description: OK
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Session not found
      summary: Execute arbitrary JavaScript
      tags:
      - Script
...
  /v1/sessions/{id}/viewport:
    get:
      operationId: getViewport
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ViewportStateResponse'
          description: OK
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Session not found
      summary: Get current viewport size and scroll offset
      tags:
      - Scrolling
servers:
- description: Internal VPC endpoint
  url: http://browser-service.internal/v1
- description: Local development
  url: http://localhost:8080/v1
...